Text now works, voice ehhh we're working on it
Text has been fixed, i honestly have no idea what i did but hey, it works
im not gonna question it. Voice does as well, kind of, AI recognizes what
i say but it starts talking to itself, well if only it talked. For some
reason the .wav file that contains the output of the generated text does
not play. So yeah, WIP